move from gitpackage to pygit2
GitPackage is impacted by the following CVE-2022-24439, ARI doesn't use the
affect method, but the present of the package is enough to raise a warning in
Quay.io vulnerability scanner.

This commit move from GitPackage to pygit2 which provides similar feature set.

Also unkile GitPython, pygit2 doesn't require the presence of the `git` binary.